`Trie::insert`, `Trie::~Trie`
The `Trie` is based on theory
"Tree Bitmap : Hardware/Software IP Lookups with Incremental Updates"
by W. Eatherton, Z. Dittia, G. Varghese.

The paper "https://cseweb.ucsd.edu/~varghese/PAPERS/ccr2004.pdf".

The stride is 5 bits. It is inferred by the following logic:
* node stores one pointer;
* pointer size on x86_64 is 8 bytes, therefore, if we are going
to store any additional data in nodes it has to be at least 8 bytes big,
otherwise we'll waste space due to alignment of 8;
* so 8 bytes is split between external and internal bit maps -
32 bits and 31 bits respectively;
* 1 bit is wasted in each node.

Thus node size is 16 bytes. Value pointers are stored contiguously with nodes.
One node can have up to 31 values, so 8 bytes is wasted on each node.

If value is trivial and is 8 bytes big, the value could be stored
in-place.

Intuitively the Trie with the configuration described above
should be optimal for the function of memory consumption, search and
implementation complexities.

Any additional sophistication will be evolved from this base. The
evolution will be benchmark-driven.